+++ docs/pages/components/button/api/button.js
@@ -40,6 +40,13 @@ export default [
                 default: '<code>false</code>'
             },
             {
+                name: '<code>expanded</code>',
+                description: 'Button will be expanded (full-width)',
+                type: 'Boolean',
+                values: '—',
+                default: '<code>false</code>'
+            },
+            {
                 name: '<code>icon-left</code>',
                 description: 'Icon name to show on the left',
                 type: 'String',
@@ -86,6 +93,11 @@ export default [
             {
                 name: '<code>click</code>',
                 description: 'Triggers on click',
+                parameters: '<code>event: $event</code>'
+            },
+            {
+                name: '<code>[any].native</code>',
+                description: 'Listen to any event using this syntax, e.g <code>mousedown.native</code>',
                 parameters: '<code>event: $event</code>'
             }
         ]