+++ docs/pages/components/checkbox/variables/checkbox.js
@@ -4,6 +4,22 @@ export default [
         default: '<code>$primary</code>'
     },
     {
+        name: '<code>$checkbox-background-color</code>',
+        default: '<code>transparent</code>'
+    },
+    {
+        name: '<code>$checkbox-border-color</code>',
+        default: '<code>$grey</code>'
+    },
+    {
+        name: '<code>$checkbox-border-radius</code>',
+        default: '<code>$radius</code>'
+    },
+    {
+        name: '<code>$checkbox-border-width</code>',
+        default: '<code>2px</code>'
+    },
+    {
         name: '<code>$checkbox-checkmark-color</code>',
         default: '<code>$primary-invert</code>'
     },