+++ docs/pages/components/loading/variables/loading.js
@@ -1,0 +1,12 @@
+export default [
+    {
+        name: '<code>$loading-background</code>',
+        description: 'The loading background color',
+        default: '<code>rgba(255,255,255,0.5)</code>'
+    },
+    {
+        name: '<code>$loading-background-legacy</code>',
+        description: 'The loading background color for non-rgba compatible browsers',
+        default: '<code>#7f7f7f</code>'
+    },
+]