+++ docs/pages/components/steps/examples/ExSimple.vue
@@ -7,7 +7,10 @@
             <b-switch v-model="showSocial"> Show Social step </b-switch>
             <b-switch v-model="isAnimated"> Animated </b-switch>
             <b-switch v-model="isStepsClickable"> Clickable Marker </b-switch>
+        </div>
+        <div class="block">
             <b-switch v-model="hasNavigation"> Navigation Buttons </b-switch>
+            <b-switch v-model="customNavigation"> Custom Navigation </b-switch>
             <b-switch v-model="isProfileSuccess"> Set <code>is-success</code> for profile </b-switch>
         </div>
         <b-field v-if="hasNavigation" grouped group-multiline>
@@ -45,6 +48,30 @@
                 <h1 class="title has-text-centered">Finish</h1>
                 Lorem ipsum dolor sit amet.
             </b-step-item>
+
+            <template
+                v-if="customNavigation"
+                slot="navigation"
+                slot-scope="{previous, next}">
+                <b-button
+                    outlined
+                    type="is-danger"
+                    icon-pack="fas"
+                    icon-left="backward"
+                    :disabled="previous.disabled"
+                    @click.prevent="previous.action">
+                    Previous
+                </b-button>
+                <b-button
+                    outlined
+                    type="is-success"
+                    icon-pack="fas"
+                    icon-right="forward"
+                    :disabled="next.disabled"
+                    @click.prevent="next.action">
+                    Next
+                </b-button>
+            </template>
         </b-steps>
     </section>
 </template>
@@ -57,6 +84,7 @@
                 showSocial: false,
                 isAnimated: true,
                 hasNavigation: true,
+                customNavigation: false,
                 prevIcon: 'chevron-left',
                 nextIcon: 'chevron-right',
                 isStepsClickable: false,