+++ docs/pages/components/toast/variables/toast.js
@@ -1,0 +1,10 @@
+export default [
+    {
+        name: '<code>$toast-border-radius</code>',
+        default: '<code>2em</code>'
+    },
+    {
+        name: '<code>$toast-opacity</code>',
+        default: '<code>0.92</code>'
+    }
+]