+++ src/components/button/Button.spec.js
@@ -1,5 +1,6 @@
 import { shallowMount, mount } from '@vue/test-utils'
 import BButton from '@components/button/Button'
+import config, {setOptions} from '@utils/config'
 
 let wrapper
 
@@ -19,7 +20,11 @@ describe('BButton', () => {
 
     it('emit a click event', () => {
         const click = jest.fn()
-        wrapper.vm.$on('click', click)
+        wrapper = shallowMount(BButton, {
+            listeners: {
+                'click': click
+            }
+        })
         wrapper.find('.button').trigger('click')
         expect(click).toHaveBeenCalledTimes(1)
     })
@@ -62,5 +67,17 @@ describe('BButton', () => {
         })
         expect(wrapper.classes()).toContain('is-large')
         expect(wrapper.contains('.icon')).toBe(true)
+    })
+
+    it('should be rounded when default config set to true', () => {
+        setOptions(Object.assign(config, {
+            defaultButtonRounded: true
+        }))
+        wrapper = mount(BButton, {
+            propsData: {
+                rounded: config.defaultButtonRounded
+            }
+        })
+        expect(wrapper.classes()).toContain('is-rounded')
     })
 })