+++ src/components/datetimepicker/index.js
@@ -1,0 +1,13 @@
+import Datetimepicker from './Datetimepicker'
+
+import { use, registerComponent } from '../../utils/plugins'
+
+const Plugin = {
+    install(Vue) {
+        registerComponent(Vue, Datetimepicker)
+    }
+}
+
+use(Plugin)
+
+export default Plugin