+++ src/components/numberinput/Numberinput.spec.js
@@ -1,37 +1,89 @@
 import { shallowMount, mount } from '@vue/test-utils'
 import BNumberinput from '@components/numberinput/Numberinput'
 
+let wrapper
+
 describe('BNumberinput', () => {
-    it('is called', () => {
-        const wrapper = shallowMount(BNumberinput)
-        expect(wrapper.name()).toBe('BNumberinput')
-        expect(wrapper.isVueInstance()).toBeTruthy()
-    })
-
-    it('renders by default', () => {
-        const wrapper = mount(BNumberinput)
-
-        expect(wrapper.contains('input')).toBeTruthy()
-        expect(wrapper.contains('button')).toBeTruthy()
-        expect(wrapper.classes()).toContain('is-grouped')
-    })
-
-    it('set controls position', () => {
-        const wrapper = mount(BNumberinput, {
-            propsData: {
-                controlsPosition: 'compact'
-            }
+    describe('Rendered', () => {
+        beforeEach(() => {
+            wrapper = mount(BNumberinput)
         })
 
-        expect(wrapper.contains('input')).toBeTruthy()
-        expect(wrapper.contains('button')).toBeTruthy()
-        expect(wrapper.classes()).toContain('has-addons')
+        it('is called', () => {
+            expect(wrapper.name()).toBe('BNumberinput')
+            expect(wrapper.isVueInstance()).toBeTruthy()
+        })
+
+        it('renders by default', () => {
+            expect(wrapper.contains('input')).toBeTruthy()
+            expect(wrapper.contains('button')).toBeTruthy()
+            expect(wrapper.classes()).toContain('is-grouped')
+        })
+
+        it('set controls position', () => {
+            wrapper.setProps({
+                controlsPosition: 'compact'
+            })
+
+            expect(wrapper.contains('input')).toBeTruthy()
+            expect(wrapper.contains('button')).toBeTruthy()
+            expect(wrapper.classes()).toContain('has-addons')
+        })
+
+        it('default value is 0', () => {
+            const input = wrapper.find('input')
+            expect(input.element.value).toBe('0')
+        })
     })
 
-    it('default value is 0', () => {
-        const wrapper = mount(BNumberinput)
+    describe('Rendered', () => {
+        beforeEach(() => {
+            wrapper = shallowMount(BNumberinput)
+        })
 
-        const input = wrapper.find('input')
-        expect(input.element.value).toBe('0')
+        it('manage prop types (number / string)', () => {
+            const min = 5
+            const max = 15
+            const step = 5
+            const stepDec = 1.5
+
+            wrapper.setProps({ min })
+            expect(wrapper.vm.minNumber).toBe(min)
+            wrapper.setProps({ min: `${min}` })
+            expect(wrapper.vm.minNumber).toBe(min)
+
+            wrapper.setProps({ max })
+            expect(wrapper.vm.maxNumber).toBe(max)
+            wrapper.setProps({ max: `${max}` })
+            expect(wrapper.vm.maxNumber).toBe(max)
+
+            wrapper.vm.newStep = step
+            expect(wrapper.vm.stepNumber).toBe(step)
+            wrapper.vm.newStep = `${step}`
+            expect(wrapper.vm.stepNumber).toBe(step)
+
+            wrapper.vm.newStep = step
+            expect(wrapper.vm.stepDecimals).toBe(0)
+            wrapper.vm.newStep = stepDec
+            expect(wrapper.vm.stepDecimals).toBe(1)
+        })
+
+        it('can increment / decrement', () => {
+            const min = 5
+            const max = 6
+            wrapper.vm.computedValue = max
+
+            wrapper.setProps({ min })
+            wrapper.vm.decrement()
+            expect(wrapper.vm.computedValue).toBe(min)
+            wrapper.vm.decrement()
+            expect(wrapper.vm.computedValue).toBe(min)
+
+            wrapper.setProps({ max })
+            wrapper.vm.increment()
+            expect(wrapper.vm.computedValue).toBe(max)
+            wrapper.vm.increment()
+            expect(wrapper.vm.computedValue).toBe(max)
+        })
     })
 })