+++ src/utils/CheckRadioMixin.js
@@ -1,0 +1,41 @@
+export default {
+    props: {
+        value: [String, Number, Boolean, Function, Object, Array],
+        nativeValue: [String, Number, Boolean, Function, Object, Array],
+        type: String,
+        disabled: Boolean,
+        required: Boolean,
+        name: String,
+        size: String
+    },
+    data() {
+        return {
+            newValue: this.value
+        }
+    },
+    computed: {
+        computedValue: {
+            get() {
+                return this.newValue
+            },
+            set(value) {
+                this.newValue = value
+                this.$emit('input', value)
+            }
+        }
+    },
+    watch: {
+        /**
+        * When v-model change, set internal value.
+        */
+        value(value) {
+            this.newValue = value
+        }
+    },
+    methods: {
+        focus() {
+            // MacOS FireFox and Safari do not focus when clicked
+            this.$refs.input.focus()
+        }
+    }
+}