+++ src/utils/config.js
@@ -40,6 +40,7 @@ let config = {
     defaultDatepickerNearbySelectableMonthDays: false,
     defaultDatepickerShowWeekNumber: false,
     defaultTrapFocus: false,
+    defaultButtonRounded: false,
 
     customIconPacks: null
 } // TODO defaultTrapFocus to true in the next breaking change